feat: Merge PersonAttachment into person__combined
OLIVIA: Two explanations/notes on these changes

1. prep__person_attachment must be included in lookup in the `job`
method so it will be available for the Count::MatchingRowsInLookup
transform. If we were only doing MergeTable on person_attachment, the
lookup would be handled automatically by the transform.

2. Comment I'm removing from the code and putting here:

This job definition clearly shows the distinction I was trying to make
in our last Migration Specialists meeting about MergeTable (for
merging in all or most of a table) vs. Merge::MultiRowLookup (for
merging in just a few fields)