feat: Add prep__filter_metadata_restriction job
I can't find where to look up the :restrictionid, :objecttype, or
:fieldtype values merged in from the FilterMetadata table.

Also this table has 235 rows, whereas the FilterMetadata table has
1011 rows. FilterMetadata contains other kind of filters (check the
parameters that contain "defaultview" or "enum" or just the name of
another table. But it also has lots of rows for fields that don't seem
like they should ever be filtered/controlled.