feat: Add jobs to compile info on dictionaries and their usage
I have no idea if this aspect of the database can be trusted or not.
That is, if ALL lookups against DictionaryItem table need to be
registered in FilterMetadata or not. Though it does seem like there
has to be SOME mechanism that the application uses to limit/filter the values
you can select for a given field, and this does seem like the place
where that happens.

If I thought it could be trusted, I might make a prep__dictionary_item
job that keeps only rows that are used. Might. It's clearly not
necessary to do so.