fix!: Return prefixed field names as Symbols
It is expected that all field names will be Symbols, and it gets
really confusing if some of them are Strings. You can't tell why some
transforms aren't working from looking at the CSV output, since the
":" in front of the header values are removed.

If any transforms have been written specifically matching on String
field names output by this transform, they will break because of this change.