feat: Fix prep__archive_identity_people -> prep__archive_identity
The lookup id for ArchiveIdentityPeople is :id (not used in this
project, but ??? in the future). Because we need to lookup on a
different id, we can't use MergeTable, which is really
inflexible (makes a lot of assumptions that this doesn't meet).

We also want to ensure (where it may actually make a difference in
future projects) that we are sorting merged values in the right order,
but not merging in the actual value of the position field from
ArchiveIdentityPeople.